Process more than one student in parallel
Use -p <int> to change the number, default is 6
Please note that the student's repos were already fetched in parallel
so if you fetch 6 students at once and there are 5 repos in that SI
you wil make 6 x 5 = 30 parallel fetches to the github API.

The agithub library has builtin throttle, so we shouldn't worry very
much.
If we hit the throttle things could actually get slower so a little
worry is healthy.